[build] add $(EXT) to tests/opaque-munit$(EXT)
This commit ensures that if you call "make EXT=.exe tests.exe", then the build
makes both tests/opaque.exe and tests/opaque-munit.exe.

I also replaced so with $(SOEXT) in the all target for consistency with how we
reference libopaque.so in the rest of the file.

I removed $(EXT) in the following:
```
test: tests/opaque-munit$(EXT)
	LD_PRELOAD=./libopaque.so ./tests/munit-opaque --fatal-failures
```
If EXT were .exe, the test command would not make sense since LD_PRELOAD is
not for Windows.