[refactor] rename sk to skS in opaque_Store1kUserRecord
to match the same variable in opaque_Register for the global server private
key. This rename prevents confusion with sk elsewhere for the shared secret
established between the user and server.

I renamed pk to pkS in opaque_Create1kRegistrationResponse to match the same
parameter in opaque_RecoverCredentials.

I added const to UserAuth's ctx since it is an input variable