[build] create a debug Makefile target
and include it in the GitHub action so that we test a debug build and a
production build whenever someone makes a pull request.

I forgot to compile a debug build the last time I made a pull request:
abbe6ececd7ea2e2d372ec8ee54affd064543939 .

This resulted in @stef having to come in and clean up:
70d567a3496a506512167a7a80b737bf19368891 .

I would have known about the problem sooner if GitHub had failed the pull
request.

https://www.gnu.org/software/make/manual/html_node/Target_002dspecific.html
pointed me to target-specific variable values.