fix incorrect SHA512 hash due to state reuse
The function `calc_preamble` ends by calling `crypto_hash_sha512_final`
on it. After this (according to the [libsodium documentation]) the
passed-in state parameter must not be re-used. However, in our case
the parameter will be re-used, because the preamble will be used
not only for the `server_mac`, but also for the `client_mac`—hence,
the SHA512 sum that was calculated before this commit was incorrect.

This has been fixed by copying the state parameter before passing
it to `crypto_hash_sha512_final`. That way, the caller of
`calc_preamble` can still use the state and the correct SHA512 sum will
be calculated.

[libsodium documentation]: https://libsodium.gitbook.io/doc/advanced/sha-2_hash_function#notes